requirements.txt
flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
markupsafe==2.1.3
requests==2.31.0
python-dotenv==1.0.0